-- implemented SetFdReuseAddress -- implemented timestamp roll-over for RTSP streams -- fixed localStreamName override when pulling in streams -- added test for SetFdReuseAddress
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@231 01689507-d990-df11-9d7d-0017a411d416

+++ sources/thelib/include/protocols/rtp/rtpheader.h
@@ -34,7 +34,7 @@
 
 typedef struct DLLEXP _RTPHeader {
 	uint32_t _flags;
-	uint32_t _timestamp;
+	uint64_t _timestamp;
 	uint32_t _ssrc;
 	//we are not interested in this yet
 	//vector<uint32_t> _csrc;
